fix previous commit (about seek)

+++ trunk/format/mpeg3.c
@@ -45,13 +45,13 @@
 	#endif
 		if (!mpeg) RAISE("IO Error: can't open file `%s': %s", filename.data(), strerror(errno));
 	}
-	\decl 0 seek (int frame);
+	\decl 0 seek (long frame);
 	\decl 0 rewind ();
 	\decl 0 bang ();
 };
 
-\def 0 seek (int frame) {
-	mpeg3_set_frame(mpeg,clip(frame,0,mpeg3_video_frames(mpeg,track)-1),track);
+\def 0 seek (long frame) {
+	mpeg3_set_frame(mpeg,clip(frame,0L,mpeg3_video_frames(mpeg,track)-1),track);
 }
 \def 0 rewind () {_0_seek(0,0,0);}
